backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
numpy>=2.1
scipy>=1.14
pandas>=2.2
httpx>=0.28
pydantic>=2.10
python-dotenv>=1.0
aiosqlite>=0.20
